decomposed SideBar.tsx
